[FIX] repair,sale_stock: display SN/lot in repair order invoices
Bug:
Invoices created for sales orders generated by a repair order didn't
show SN/lot number for products that are tracked when activating
"Display Lots & Serial Numbers on Invoices" setting. This was because
the method generating the lots for the report only included stock move
lines that had `customer` in its source or destination location usage,
which is not the case in the move lines of a repair order.

Fix:
Exclude repair move lines from this check.

Task-3848611

closes odoo/odoo#167187

X-original-commit: 2d660278f92d2bd42a82e52c0d06d8f6984832a0
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>